# WARNING: head commit changed in the meantime
Merge branch 'feature-1.5.0-backend' of
https://gitlab.ak12.cc/dx-server/dx-common-facades.git into
feature-1.5.0-backend

+++ src/main/java/com/dx/facade/member/param/AddRebateRateTemplateMappingParamDTO.java
@@ -38,7 +38,7 @@ public class AddRebateRateTemplateMappingParamDTO implements Serializable {
     private Integer userType;
     
     @ApiModelProperty("总代Id")
-    private Integer topProxyId;
+    private Long topProxyId;
     
     @ApiModelProperty("创建人")
     private String createdBy;
